+++ crates/objc2/src/__framework_prelude.rs
@@ -14,16 +14,13 @@ pub use crate::encode::{Encode, Encoding, RefEncode};
 pub use crate::ffi::{NSInteger, NSIntegerMax, NSUInteger, NSUIntegerMax};
 pub use crate::rc::{Allocated, DefaultRetained, Retained};
 pub use crate::runtime::{
-    AnyClass, AnyObject, Bool, NSObject, NSObjectProtocol, ProtocolObject, Sel,
+    AnyClass, AnyObject, AnyProtocol, Bool, NSObject, NSObjectProtocol, ProtocolObject, Sel,
 };
 pub use crate::{
     extern_category, extern_class, extern_methods, extern_protocol, ClassType, MainThreadMarker,
     MainThreadOnly, Message, ProtocolType,
 };
 
-// TODO
-pub type AnyProtocol = AnyObject;
 pub type TodoFunction = *const c_void;
-pub type TodoClass = AnyObject;
 pub type TodoProtocols = AnyObject;
 pub type IMP = Option<crate::runtime::Imp>;