+++ crates/objc2/tests/protocol_shared_retain_count.rs
@@ -1,0 +1,24 @@
+//! Test that AnyProtocol objects have a shared retain count.
+//!
+//! Separate test because this will likely not work if other tests are running
+//! at the same time.
+
+use objc2::runtime::{AnyObject, NSObject, NSObjectProtocol};
+use objc2::{Message, ProtocolType};
+
+#[test]
+fn protocol_has_shared_retain_count() {
+    let obj: &AnyObject = <dyn NSObjectProtocol>::protocol().unwrap().as_ref();
+    let obj = obj.downcast_ref::<NSObject>().unwrap();
+
+    assert_eq!(obj.retainCount(), 1);
+    let obj2 = obj.retain();
+    assert_eq!(obj.retainCount(), 2);
+    drop(obj2);
+    assert_eq!(obj.retainCount(), 1);
+
+    let obj2: &AnyObject = <dyn NSObjectProtocol>::protocol().unwrap().as_ref();
+    assert_eq!(obj.retainCount(), 1);
+    let _obj2 = obj2.retain();
+    assert_eq!(obj.retainCount(), 2);
+}